Added error message capture during export

+++ src/python/scenario_model.py
@@ -857,17 +857,20 @@ class ImageProjectModel:
       return nfile,im
 
     def export(self, location):
-      self.G.create_archive(location)
+      path,errors = self.G.create_archive(location)
+      return errors
 
     def exporttos3(self, location):
       import boto3
-      path = self.G.create_archive(tempfile.gettempdir())
-      s3 = boto3.client('s3','us-east-1')
-      BUCKET = location.split('/')[0].strip()
-      DIR= location.split('/')[1].strip()
-      print 'Upload to s3://' + BUCKET + '/' + DIR + '/' + os.path.split(path)[1] 
-      s3.upload_file(path, BUCKET, DIR + '/' + os.path.split(path)[1])
-      os.remove(path)
+      path,errors = self.G.create_archive(tempfile.gettempdir())
+      if len(errors) == 0:
+        s3 = boto3.client('s3','us-east-1')
+        BUCKET = location.split('/')[0].strip()
+        DIR= location.split('/')[1].strip()
+        print 'Upload to s3://' + BUCKET + '/' + DIR + '/' + os.path.split(path)[1] 
+        s3.upload_file(path, BUCKET, DIR + '/' + os.path.split(path)[1])
+        os.remove(path)
+      return errors
 
     def export_path(self, location):
       if self.end is None and self.start is not None:
